修改bug

+++ 30000day/1.HomePage/2.Person/controller/PersonViewController.m
@@ -128,7 +128,8 @@
 - (void)getMyFriends {
 
     [MTProgressHUD showHUD:[UIApplication sharedApplication].keyWindow];
-    [STDataHandler getMyFriendsWithUserId:[NSString stringWithFormat:@"%@",STUserAccountHandler.userProfile.userId] order:[NSString stringWithFormat:@"%ld",self.sortTab] success:^(NSMutableArray *dataArray) {
+    
+    [STDataHandler getMyFriendsWithUserId:[NSString stringWithFormat:@"%@",STUserAccountHandler.userProfile.userId] order:[NSString stringWithFormat:@"%d",(int)self.sortTab] success:^(NSMutableArray *dataArray) {
         
         dispatch_async(dispatch_get_main_queue(), ^{
